requirements.txt
# Core
numpy
requests
python-dotenv
tqdm

# Document parsing
docling

# Vector DB / retrieval
faiss-cpu # faiss-gpu если есть GPU
sentence-transformers

# LLM / LM Studio
lmstudio

# Documents handling
pandas

# Optional faster torch for reranker (install matching CUDA if you have GPU)
torch
